Working nix-shell
See the diffs; removing the extra dirs options from the cabal file and
adding a dedicated shell.nix seems to do the trick. Why that is, is
still confusing to me.

I believe mkShell, where zlib is repeated as a buildInput, sets the
paths in the shell so that cabal can find it. But this is just a guess
and further reading is needed.