Test with extra-libraries in cabal file
Adding an extra-libraries option set to zlib seems to cause build
failures.
These changes show how to make the issue happen.
After this, we need to run the shell:

```
nix-shell --attr avahi-dbus.env release.nix
```

Following with doing `cabal repl` fails with:

```
[nix-shell:~/code/homebot/rpi/avahi-dbus]$ cabal repl
Resolving dependencies...
Build profile: -w ghc-8.6.5 -O1
In order, the following will be built (use -v for more details):
 - avahi-dbus-0.1.0.0 (lib) (configuration changed)
Configuring library for avahi-dbus-0.1.0.0..
cabal: Missing dependency on a foreign library:
* Missing (or bad) C library: zlib
This problem can usually be solved by installing the system package that
provides this library (you may need the "-dev" version). If the library is
already installed but in a non-standard location then you can use the flags
--extra-include-dirs= and --extra-lib-dirs= to specify where it is.If the
library file does exist, it may contain errors that are caught by the C
compiler at the preprocessing stage. In this case you can re-run configure
with the verbosity flag -v3 to see the error messages.
```

The failures go away when the extra-libraries option in the cabal file
is commented out.

Why is this?